+++ TuEventoBackend/tu-evento/src/main/java/com/capysoft/tuevento/modules/security/application/port/in/RejectOrganizerRequestPort.java
@@ -1,0 +1,5 @@
+package com.capysoft.tuevento.modules.security.application.port.in;
+
+public interface RejectOrganizerRequestPort {
+    void reject(Integer organizerPetitionId);
+}